Portail: compose les blocs indépendants sans casser le dialogue historique
