--- .github/workflows/noe063-consolidation-helper.yml
@@ -1,94 +1,0 @@
-name: Noe063 consolidation helper
-
-on:
-  push:
-    branches: [feature/noe-063-portail-consolide]
-    paths: [.github/workflows/noe063-consolidation-helper.yml]
-  pull_request:
-    branches: [master]
-
-permissions:
-  contents: write
-
-jobs:
-  patch:
-    if: github.event_name == 'pull_request'
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v7
-        with:
-          ref: feature/noe-063-portail-consolide
-          fetch-depth: 0
-
-      - name: Raccorder les blocs enrichis
-        shell: python
-        run: |
-          from pathlib import Path
-          import re
-
-          dlg_path = Path("noethys/Dlg/DLG_Saisie_portail_bloc.py")
-          dlg = dlg_path.read_text(encoding="utf-8")
-
-          marker = "from Ctrl import CTRL_Editeur_email\n"
-          assert marker in dlg
-          dlg = dlg.replace(marker, marker + "from Ctrl import CTRL_Portail_contenu_externe\nfrom Ctrl import CTRL_Portail_tarifs\nfrom Utils import UTILS_Portail_blocs\n", 1)
-
-          ligne_texte = "            (_(\"bloc_texte\"), _(u\"Texte\"), PAGE_Texte(self), wx.Bitmap(Chemins.GetStaticPath('Images/32x32/Texte_bloc.png'), wx.BITMAP_TYPE_PNG)),\n"
-          assert ligne_texte in dlg
-          ajouts = (
-              "            (UTILS_Portail_blocs.CODE_CONTENU_EXTERNE, _(u\"Contenu externe\"), CTRL_Portail_contenu_externe.CTRL(self), wx.Bitmap(Chemins.GetStaticPath('Images/32x32/Apercu.png'), wx.BITMAP_TYPE_PNG)),\n"
-              "            (UTILS_Portail_blocs.CODE_TARIFS, _(u\"Tarifs Noethys\"), CTRL_Portail_tarifs.CTRL(self), wx.Bitmap(Chemins.GetStaticPath('Images/32x32/Euro.png'), wx.BITMAP_TYPE_PNG)),\n"
-          )
-          dlg = dlg.replace(ligne_texte, ligne_texte + ajouts, 1)
-
-          ancien = '''    def GetParametres(self):
-        dictParametres = self.GetPageActive().GetParametres()
-        dictParametres["categorie"] = self.GetCodePageActive()
-        return dictParametres
-
-    def SetParametres(self, dictParametres={}):
-        self.SetPageByCode(dictParametres["categorie"])
-        self.GetPageActive().SetParametres(dictParametres)'''
-          nouveau = '''    def GetParametres(self):
-        dictParametres = self.GetPageActive().GetParametres()
-        code = self.GetCodePageActive()
-        dictParametres["categorie"] = UTILS_Portail_blocs.categorie_persistante(code)
-        return dictParametres
-
-    def SetParametres(self, dictParametres={}):
-        code = UTILS_Portail_blocs.detecter_code(dictParametres)
-        self.SetPageByCode(code)
-        self.GetPageActive().SetParametres(dictParametres)'''
-          assert ancien in dlg
-          dlg = dlg.replace(ancien, nouveau, 1)
-          dlg_path.write_text(dlg, encoding="utf-8")
-
-          serveur_path = Path("noethys/Ctrl/CTRL_Portail_serveur.py")
-          serveur = serveur_path.read_text(encoding="utf-8")
-          marker = "from Utils import UTILS_Portail_synchro\n"
-          assert marker in serveur
-          serveur = serveur.replace(marker, marker + "from Utils import UTILS_Portail_tarifs_synchro\n", 1)
-
-          cible = '                        synchro = UTILS_Portail_synchro.Synchro(log=self.parent)\n'
-          assert cible in serveur
-          preparation = '''                        # Régénère les tarifs avant l'export Connecthys historique.
-                        # Une erreur conserve le dernier rendu et ne bloque pas la synchro.
-                        try:
-                            UTILS_Portail_tarifs_synchro.preparer_avant_synchro(log=self.parent)
-                        except Exception as err:
-                            self.parent.EcritLog(_(u"[AVERTISSEMENT] Actualisation des tarifs portail impossible : %s") % err)
-
-'''
-          serveur = serveur.replace(cible, preparation + cible, 1)
-          serveur_path.write_text(serveur, encoding="utf-8")
-
-          Path(".github/workflows/noe063-consolidation-helper.yml").unlink()
-
-      - name: Commit de consolidation
-        shell: bash
-        run: |
-          git config user.name "github-actions[bot]"
-          git config user.email "41898282+github-actions[bot]@users.noreply.github.com"
-          git add -A
-          git commit -m "Portail: raccorde les blocs indépendants à Connecthys"
-          git push origin HEAD:feature/noe-063-portail-consolide
